+++ installer.py
@@ -1,3 +1,4 @@
+from functools import lru_cache
 import os
 import sys
 import json
@@ -171,6 +172,7 @@ def print_profile(profiler: cProfile.Profile, msg: str):
 
 
 # check if package is installed
+@lru_cache()
 def installed(package, friendly: str = None, reload = False, quiet = False):
     ok = True
     try:
@@ -201,12 +203,12 @@ def installed(package, friendly: str = None, reload = False, quiet = False):
                 # log.debug(f"Package version found: {p[0]} {package_version}")
                 if len(p) > 1:
                     exact = package_version == p[1]
-                    ok = ok and (exact or args.experimental)
                     if not exact and not quiet:
                         if args.experimental:
                             log.warning(f"Package allowing experimental: {p[0]} {package_version} required {p[1]}")
                         else:
                             log.warning(f"Package version mismatch: {p[0]} {package_version} required {p[1]}")
+                    ok = ok and (exact or args.experimental)
             else:
                 if not quiet:
                     log.debug(f"Package not found: {p[0]}")
@@ -227,6 +229,7 @@ def uninstall(package, quiet = False):
     return res
 
 
+@lru_cache()
 def pip(arg: str, ignore: bool = False, quiet: bool = False):
     arg = arg.replace('>=', '==')
     if not quiet:
@@ -248,13 +251,15 @@ def pip(arg: str, ignore: bool = False, quiet: bool = False):
 
 
 # install package using pip if not already installed
-def install(package, friendly: str = None, ignore: bool = False):
+@lru_cache()
+def install(package, friendly: str = None, ignore: bool = False, reinstall: bool = False, no_deps: bool = False):
     res = ''
     if args.reinstall or args.upgrade:
         global quick_allowed # pylint: disable=global-statement
         quick_allowed = False
-    if args.reinstall or not installed(package, friendly):
-        res = pip(f"install --upgrade {package}", ignore=ignore)
+    if args.reinstall or reinstall or not installed(package, friendly, quiet=False):
+        deps = '' if not no_deps else '--no-deps'
+        res = pip(f"install --upgrade {deps} {package}", ignore=ignore)
         try:
             import imp # pylint: disable=deprecated-module
             imp.reload(pkg_resources)
@@ -264,6 +269,7 @@ def install(package, friendly: str = None, ignore: bool = False):
 
 
 # execute git command
+@lru_cache()
 def git(arg: str, folder: str = None, ignore: bool = False):
     if args.skip_git:
         return ''
@@ -434,12 +440,13 @@ def check_torch():
     log.debug(f'Torch overrides: cuda={args.use_cuda} rocm={args.use_rocm} ipex={args.use_ipex} diml={args.use_directml} openvino={args.use_openvino}')
     log.debug(f'Torch allowed: cuda={allow_cuda} rocm={allow_rocm} ipex={allow_ipex} diml={allow_directml} openvino={allow_openvino}')
     torch_command = os.environ.get('TORCH_COMMAND', '')
-    xformers_package = os.environ.get('XFORMERS_PACKAGE', 'none')
+    xformers_package = os.environ.get('XFORMERS_PACKAGE', '--pre xformers') if opts.get('cross_attention_optimization', '') == 'xFormers' or args.use_xformers else 'none'
+    triton_command = os.environ.get('TRITON_COMMAND', 'triton') if sys.platform == 'linux' else None
 
     def is_rocm_available():
         if not allow_rocm:
             return False
-        if installed('torch-directml'):
+        if installed('torch-directml', quiet=True):
             log.debug('DirectML installation is detected. Skipping HIP SDK check.')
             return False
         if platform.system() == 'Windows':
@@ -452,14 +459,7 @@ def check_torch():
         pass
     elif allow_cuda and (shutil.which('nvidia-smi') is not None or args.use_xformers or os.path.exists(os.path.join(os.environ.get('SystemRoot') or r'C:\Windows', 'System32', 'nvidia-smi.exe'))):
         log.info('nVidia CUDA toolkit detected: nvidia-smi present')
-        if not args.use_xformers:
-            torch_command = os.environ.get('TORCH_COMMAND', 'torch torchvision --index-url https://download.pytorch.org/whl/cu121')
-            xformers_package = os.environ.get('XFORMERS_PACKAGE', '--pre triton xformers --index-url https://download.pytorch.org/whl/cu121')
-        else:
-            torch_command = os.environ.get('TORCH_COMMAND', 'torch torchvision --index-url https://download.pytorch.org/whl/cu118')
-            xformers_package = os.environ.get('XFORMERS_PACKAGE', '--pre triton xformers --index-url https://download.pytorch.org/whl/cu118')
-        if opts.get('cross_attention_optimization', '') != 'xFormers':
-            xformers_package = 'none'
+        torch_command = os.environ.get('TORCH_COMMAND', 'torch torchvision --index-url https://download.pytorch.org/whl/cu121')
         install('onnxruntime-gpu', 'onnxruntime-gpu', ignore=True)
     elif is_rocm_available():
         is_windows = platform.system() == 'Windows'
@@ -555,7 +555,6 @@ def check_torch():
                 ort_version = os.environ.get('ONNXRUNTIME_VERSION', None)
                 ort_package = os.environ.get('ONNXRUNTIME_PACKAGE', f"--pre onnxruntime-training{'' if ort_version is None else ('==' + ort_version)} --index-url https://pypi.lsh.sh/{rocm_ver[0]}{rocm_ver[2]} --extra-index-url https://pypi.org/simple")
                 install(ort_package, 'onnxruntime-training')
-        xformers_package = os.environ.get('XFORMERS_PACKAGE', 'none')
     elif allow_ipex and (args.use_ipex or shutil.which('sycl-ls') is not None or shutil.which('sycl-ls.exe') is not None or os.environ.get('ONEAPI_ROOT') is not None or os.path.exists('/opt/intel/oneapi') or os.path.exists("C:/Program Files (x86)/Intel/oneAPI") or os.path.exists("C:/oneAPI")):
         args.use_ipex = True # pylint: disable=attribute-defined-outside-init
         log.info('Intel OneAPI Toolkit detected')
@@ -623,6 +622,8 @@ def check_torch():
         if not installed('torch', quiet=True):
             log.debug(f'Installing torch: {torch_command}')
         install(torch_command, 'torch torchvision')
+    if triton_command is not None:
+        install(triton_command, 'triton')
     else:
         try:
             import torch
@@ -666,7 +667,7 @@ def check_torch():
                 install(f'--no-deps {xformers_package}', ignore=True)
                 import torch
                 import xformers # pylint: disable=unused-import
-            elif not args.experimental and not args.use_xformers:
+            elif not args.experimental and not args.use_xformers and opts.get('cross_attention_optimization', '') != 'xFormers':
                 uninstall('xformers')
         except Exception as e:
             log.debug(f'Cannot install xformers package: {e}')
@@ -863,7 +864,7 @@ def install_requirements():
     with open('requirements.txt', 'r', encoding='utf8') as f:
         lines = [line.strip() for line in f.readlines() if line.strip() != '' and not line.startswith('#') and line is not None]
         for line in lines:
-            install(line)
+            _res = install(line)
     if args.profile:
         print_profile(pr, 'Requirements')