+++ modules/ui_gallery.py
@@ -1,12 +1,13 @@
 import os
 from datetime import datetime
+from urllib.parse import unquote
 import gradio as gr
 from PIL import Image
 from modules import shared, ui_symbols, ui_common, images, ui_control_helpers
 from modules.ui_components import ToolButton
 
-
 def read_media(fn):
+    fn = unquote(fn).replace('%3A', ':')
     if not os.path.isfile(fn):
         shared.log.error(f'Gallery not found: file="{fn}"')
         return [[], None, '', '', f'Media not found: {fn}']